mon: Monitor: cleanup timecheck state before changing the monitor state
This must always be done prior to start an election, or whenever we
bootstrap or reset the monitor.

We know that by cleaning up the timecheck state both on bootstrap and
reset we are duplicating some effort whenever bootstrap is called. But
the truth is that reset is called often in some contexts in which
bootstrapping is not required, and we also need to cleanup the timecheck
state in those occasions.

Fixes: #3695

Signed-off-by: Joao Eduardo Luis <joao.luis@inktank.com>